dinD

+++ Programs/One-Wire-Bus/Inc/Input.h
@@ -37,7 +37,7 @@ int temperatur_lesen(uint8_t *sensor_rom, float *temperatur);
 
 int OW_search(uint8_t *ROM_NUM,uint8_t *LastDeviceFlag,uint8_t *Lastdiscrepans);
 
-int OWFirst(int LastDiscrepancy,int LastDeviceFlag);
+int OWFirst( uint8_t LastDiscrepancy,uint8_t LastDeviceFlag);
 int OW_searcht(uint8_t *ROM_NUM,uint8_t *LastDeviceFlag,uint8_t *Lastdiscrepans);
 
 #endif
